Reference

Contents

Reference

Contents

Index

Index

MPCC.AbstractMPCCModel

MPCC.MPCCCounters

MPCC.MPCCModelMeta

MPCC.NLMPCC

LinearOperators.reset!

LinearOperators.reset!

MPCC.complementarity_constrained

MPCC.consG

MPCC.consG!

MPCC.consH

MPCC.consH!

MPCC.decrement_cc!

MPCC.hGprod

MPCC.hGprod!

MPCC.hGprod!

MPCC.hHprod

MPCC.hHprod!

MPCC.hHprod!

MPCC.hessG

MPCC.hessG_coord

MPCC.hessG_coord!

MPCC.hessG_op

MPCC.hessG_op!

MPCC.hessG_structure

MPCC.hessG_structure!

MPCC.hessH

MPCC.hessH_coord

MPCC.hessH_coord!

MPCC.hessH_op

MPCC.hessH_structure

MPCC.hessH_structure!

MPCC.increment_cc!

MPCC.jGprod

MPCC.jGprod!

MPCC.jGprod!

MPCC.jGtprod

MPCC.jGtprod!

MPCC.jGtprod!

MPCC.jHprod

MPCC.jHprod!

MPCC.jHprod!

MPCC.jHtprod

MPCC.jHtprod!

MPCC.jHtprod!

MPCC.jacG

MPCC.jacG_coord

MPCC.jacG_coord!

MPCC.jacG_op

MPCC.jacG_op!

MPCC.jacG_structure

MPCC.jacG_structure!

MPCC.jacH

MPCC.jacH_coord

MPCC.jacH_coord!

MPCC.jacH_op

MPCC.jacH_op!

MPCC.jacH_structure

MPCC.jacH_structure!

MPCC.neval_consG

MPCC.neval_consH

MPCC.neval_hGprod

MPCC.neval_hHprod

MPCC.neval_hessG

MPCC.neval_hessH

MPCC.neval_jGprod

MPCC.neval_jGtprod

MPCC.neval_jHprod

MPCC.neval_jHtprod

MPCC.neval_jacG

MPCC.neval_jacH

MPCC.viol

MPCC.viol

MPCC.viol!

MPCC.AbstractMPCCModel — Type

Base type for an optimization model with degenerate constraints.

min f(x)
l <= x <= u
lb <= c(x) <= ub
0 <= G(x) _|_ H(x) >= 0

MPCC.MPCCCounters — Type

MPCCCounters

Initialization: MPCCCounters()

MPCC.MPCCModelMeta — Type

A composite type that represents the main features of the optimization problem

optimize obj(x) subject to lvar ≤ x ≤ uvar lcon ≤ cons(x) ≤ ucon lcc ≤ G(x) | H(x) >= ucc

where x is an nvar-dimensional vector, obj is the real-valued objective function, cons is the vector-valued constraint function, optimize is either "minimize" or "maximize".

Here, lvar, uvar, lcon and ucon are vectors. Some of their components may be infinite to indicate that the corresponding bound or general constraint is not present.

MPCC.NLMPCC — Type

Convert an MPCCModel to an NLPModels as follows. Definit le type NLMPCC : min f(x) s.t. l <= x <= u lcon(tb) <= cnl(x) <= ucon

with

cnl(x) := c(x),G(x),H(x),G(x).*H(x)

LinearOperators.reset! — Method

reset!(nlp)

Reset evaluation count in nlp

LinearOperators.reset! — Method

reset!(counters)

Reset evaluation counters

MPCC.complementarity_constrained — Method

complementarity_constrained(nlp)
complementarity_constrained(meta)

Returns whether the problem's constraints are all inequalities. Unconstrained problems return true.

MPCC.consG! — Method

Evaluate $G(x)$, the constraints at x.

MPCC.consG — Method

Evaluate $G(x)$, the constraints at x.

MPCC.consH! — Method

Evaluate $H(x)$, the constraints at x.

MPCC.consH — Method

Evaluate $G(x)$, the constraints at x.

MPCC.decrement_cc! — Method

decrement!(nlp, s)

Decrement counter s of problem nlp.

MPCC.hGprod! — Method

Hv = hGprod!(nlp, x, v, Hv; objweight=1.0) Evaluate the product of the objective Hessian at x with the vector v in place, with objective function scaled by `objweight`, where the objective Hessian is OBJECTIVE_HESSIAN.

MPCC.hGprod! — Method

Hv = hGprod!(nlp, x, y, v, Hv; objweight=1.0) Evaluate the product of the Lagrangian Hessian at (x,y) with the vector v in place, with objective function scaled by `objweight`, where the Lagrangian Hessian is LAGRANGIAN_HESSIAN.

MPCC.hGprod — Method

Hv = hGprod(nlp, x, y, v)

Evaluate the product of the G Hessian at (x,y) with the vector v.

MPCC.hHprod! — Method

Hv = hHprod!(nlp, x, v, Hv; objweight=1.0) Evaluate the product of the objective Hessian at x with the vector v in place, with objective function scaled by `objweight`, where the objective Hessian is OBJECTIVE_HESSIAN.

MPCC.hHprod! — Method

Hv = hHprod!(nlp, x, y, v, Hv; objweight=1.0) Evaluate the product of the Lagrangian Hessian at (x,y) with the vector v in place, with objective function scaled by `objweight`, where the Lagrangian Hessian is LAGRANGIAN_HESSIAN.

MPCC.hHprod — Method

Hv = hHprod(nlp, x, y, v)

Evaluate the product of the G Hessian at (x,y) with the vector v.

MPCC.hessG — Method

Hx = hessG(nlp, x, y)

Evaluate the Lagrangian Hessian at (x,y) as a sparse matrix, with objective function scaled by obj_weight, i.e., LAGRANGIAN_HESSIAN. Only the lower triangle is returned.

MPCC.hessG_coord! — Method

vals = hess_coord!(nlp, x, y, vals)

Evaluate the Lagrangian Hessian at (x,y) in sparse coordinate format, with objective function scaled by obj_weight, i.e., LAGRANGIAN_HESSIAN, rewriting vals. Only the lower triangle is returned.

MPCC.hessG_coord — Method

vals = hessG_coord(nlp, x, y)

Evaluate the Lagrangian Hessian at (x,y) in sparse coordinate format, with objective function scaled by obj_weight, i.e., LAGRANGIAN_HESSIAN. Only the lower triangle is returned.

MPCC.hessG_op! — Method

H = hessG_op!(nlp, x, y, Hv; obj_weight=1.0)

Return the Lagrangian Hessian at (x,y) with objective function scaled by obj_weight as a linear operator, and storing the result on Hv. The resulting object may be used as if it were a matrix, e.g., w = H * v. The vector Hv is used as preallocated storage for the operation. The linear operator H represents LAGRANGIAN_HESSIAN.

MPCC.hessG_op — Method

H = hessG_op(nlp, x, y; obj_weight=1.0)

Return the Lagrangian Hessian at (x,y) with objective function scaled by obj_weight as a linear operator. The resulting object may be used as if it were a matrix, e.g., H * v. The linear operator H represents LAGRANGIAN_HESSIAN.

MPCC.hessG_structure! — Method

hessG_structure!(nlp, rows, cols)

Return the structure of the Lagrangian Hessian in sparse coordinate format in place.

MPCC.hessG_structure — Method

(rows,cols) = hessG_structure(nlp)

Return the structure of the Lagrangian Hessian in sparse coordinate format.

MPCC.hessH — Method

Hx = hessH(nlp, x, y)

Evaluate the Lagrangian Hessian at (x,y) as a sparse matrix, with objective function scaled by obj_weight, i.e., LAGRANGIAN_HESSIAN. Only the lower triangle is returned.

MPCC.hessH_coord! — Method

vals = hessH_coord!(nlp, x, y, vals)

Evaluate the Lagrangian Hessian at (x,y) in sparse coordinate format, with objective function scaled by obj_weight, i.e., LAGRANGIAN_HESSIAN, rewriting vals. Only the lower triangle is returned.

MPCC.hessH_coord — Method

vals = hessG_coord(nlp, x, y)

Evaluate the Lagrangian Hessian at (x,y) in sparse coordinate format, with objective function scaled by obj_weight, i.e., LAGRANGIAN_HESSIAN. Only the lower triangle is returned.

MPCC.hessH_op — Method

H = hessH_op(nlp, x, y; obj_weight=1.0)

Return the Lagrangian Hessian at (x,y) with objective function scaled by obj_weight as a linear operator. The resulting object may be used as if it were a matrix, e.g., H * v. The linear operator H represents LAGRANGIAN_HESSIAN.

MPCC.hessH_structure! — Method

hessH_structure!(nlp, rows, cols)

Return the structure of the Lagrangian Hessian in sparse coordinate format in place.

MPCC.hessH_structure — Method

(rows,cols) = hessG_structure(nlp)

Return the structure of the Lagrangian Hessian in sparse coordinate format.

MPCC.increment_cc! — Method

increment_cc!(nlp, s)

Increment counter s of problem nlp.

MPCC.jGprod! — Method

JGv = jGprod!(nlp, x, v, Jv) Evaluate $∇G(x)v$, the Jacobian-vector product at x in place.

MPCC.jGprod! — Method

JGv = jGprod!(nlp, x, v, Jv) Evaluate $∇G(x)v$, the Jacobian-vector product at x in place.

MPCC.jGprod — Method

JGv = jGprod(nlp, x, v)

Evaluate $∇G(x)v$, the Jacobian-vector product at x.

MPCC.jGtprod! — Method

JGtv = jGtprod!(nlp, x, v, Jtv) Evaluate $∇G(x)^Tv$, the transposed-Jacobian-vector product at x in place.

MPCC.jGtprod! — Method

Jtv = jtprodG(nlp, x, v, Jtv)

Evaluate $∇G(x)^Tv$, the transposed-Jacobian-vector product at x

MPCC.jGtprod — Method

JGtv = jGtprod(nlp, x, v, Jtv)

Evaluate $∇G(x)^Tv$, the transposed-Jacobian-vector product at x.

MPCC.jHprod! — Method

JHv = jHprod!(nlp, x, v, Jv) Evaluate $∇H(x)v$, the Jacobian-vector product at x in place.

MPCC.jHprod! — Method

JHv = jHprod!(nlp, x, v, Jv) Evaluate $∇H(x)v$, the Jacobian-vector product at x in place.

MPCC.jHprod — Method

JHv = jHprod(nlp, x, v)

Evaluate $∇H(x)v$, the Jacobian-vector product at x.

MPCC.jHtprod! — Method

JHtv = jHtprod!(nlp, x, v, Jtv) Evaluate $∇H(x)^Tv$, the transposed-Jacobian-vector product at x in place.

MPCC.jHtprod! — Method

Jtv = jtprodH(nlp, x, v, Jtv)

Evaluate $∇H(x)^Tv$, the transposed-Jacobian-vector product at x

MPCC.jHtprod — Method

JHtv = jHtprod(nlp, x, v, Jtv)

Evaluate $∇H(x)^Tv$, the transposed-Jacobian-vector product at x.

MPCC.jacG — Method

Jx = jacG(nlp, x)

Evaluate $∇G(x)$, the constraint's Jacobian at x as a sparse matrix.

MPCC.jacG_coord! — Method

vals = jacG_coord!(nlp, x, vals)

Evaluate $∇G(x)$, the constraint's Jacobian at x in sparse coordinate format, rewriting vals.

MPCC.jacG_coord — Method

vals = jacG_coord(nlp, x)

Evaluate $∇G(x)$, the constraint's Jacobian at x in sparse coordinate format.

MPCC.jacG_op! — Method

J = jacG_op!(nlp, x, Jv, Jtv)

Return the Jacobian at x as a linear operator. The resulting object may be used as if it were a matrix, e.g., J * v or J' * v. The values Jv and Jtv are used as preallocated storage for the operations.

MPCC.jacG_op — Method

J = jacG_op(nlp, x)

Return the Jacobian at x as a linear operator. The resulting object may be used as if it were a matrix, e.g., J * v or J' * v.

MPCC.jacG_structure! — Method

jacG_structure!(nlp, rows, cols)

Return the structure of the constraint's Jacobian in sparse coordinate format in place.

MPCC.jacG_structure — Method

(rows,cols) = jacG_structure(nlp)

Return the structure of the constraint's Jacobian in sparse coordinate format.

MPCC.jacH — Method

Jx = jacH(nlp, x)

Evaluate $∇H(x)$, the constraint's Jacobian at x as a sparse matrix.

MPCC.jacH_coord! — Method

vals = jacH_coord!(nlp, x, vals)

Evaluate $∇H(x)$, the constraint's Jacobian at x in sparse coordinate format, rewriting vals.

MPCC.jacH_coord — Method

vals = jacH_coord(nlp, x)

Evaluate $∇H(x)$, the constraint's Jacobian at x in sparse coordinate format.

MPCC.jacH_op! — Method

J = jacH_op!(nlp, x, Jv, Jtv)

Return the Jacobian at x as a linear operator. The resulting object may be used as if it were a matrix, e.g., J * v or J' * v. The values Jv and Jtv are used as preallocated storage for the operations.

MPCC.jacH_op — Method

J = jacH_op(nlp, x)

Return the Jacobian at x as a linear operator. The resulting object may be used as if it were a matrix, e.g., J * v or J' * v.

MPCC.jacH_structure! — Method

jacH_structure!(nlp, rows, cols)

Return the structure of the constraint's Jacobian in sparse coordinate format in place.

MPCC.jacH_structure — Method

(rows,cols) = jacH_structure(nlp)

Return the structure of the constraint's Jacobian in sparse coordinate format.

MPCC.neval_consG — Method

neval_consG(nlp)

Get the number of consG evaluations.

MPCC.neval_consH — Method

neval_consH(nlp)

Get the number of consH evaluations.

MPCC.neval_hGprod — Method

neval_hGprod(nlp)

Get the number of hGprod evaluations.

MPCC.neval_hHprod — Method

neval_hHprod(nlp)

Get the number of hHprod evaluations.

MPCC.neval_hessG — Method

neval_hessG(nlp)

Get the number of hessG evaluations.

MPCC.neval_hessH — Method

neval_hessH(nlp)

Get the number of hessH evaluations.

MPCC.neval_jGprod — Method

neval_jGprod(nlp)

Get the number of jGprod evaluations.

MPCC.neval_jGtprod — Method

neval_jGtprod(nlp)

Get the number of jGtprod evaluations.

MPCC.neval_jHprod — Method

neval_jHprod(nlp)

Get the number of jHprod evaluations.

MPCC.neval_jHtprod — Method

neval_jHtprod(nlp)

Get the number of jHtprod evaluations.

MPCC.neval_jacG — Method

neval_jacG(nlp)

Get the number of jacG evaluations.

MPCC.neval_jacH — Method

neval_jacH(nlp)

Get the number of jacH evaluations.

MPCC.viol! — Method

c = viol!(nlp, x, c)
Return the vector of the constraints
lx <= x <= ux
lc <= c(x) <= uc,
lccG <= G(x),
lccH <= H(x),
G(x) .* H(x) <= 0

MPCC.viol — Method

c = viol(nlp, x)

Evaluate $c(x)$, the constraints at x.

MPCC.viol — Method

Return the violation of the constraints lb <= x <= ub, lc <= c(x) <= uc, lccG <= G(x), lccH <= H(x), G(x).*H(x) <= 0.