MPCC

Types

MPCC.MPCCModelMeta — Type

A composite type that represents the main features of the optimization problem

optimize obj(x) subject to lvar ≤ x ≤ uvar lcon ≤ cons(x) ≤ ucon lcc ≤ G(x) | H(x) >= ucc

where x is an nvar-dimensional vector, obj is the real-valued objective function, cons is the vector-valued constraint function, optimize is either "minimize" or "maximize".

Here, lvar, uvar, lcon and ucon are vectors. Some of their components may be infinite to indicate that the corresponding bound or general constraint is not present.

MPCC.MPCCCounters — Type

MPCCCounters

Initialization: MPCCCounters()

Missing docstring.

Missing docstring for MPCC.ADMPCCModel. Check Documenter's build log for details.

Missing docstring.

Missing docstring for MPCC.BPMPCCModel. Check Documenter's build log for details.

Missing docstring.

Missing docstring for MPCC.MPCCAmpl. Check Documenter's build log for details.

Missing docstring.

Missing docstring for MPCC.MPCCNLPs. Check Documenter's build log for details.

MPCC.NLMPCC — Type

Convert an MPCCModel to an NLPModels as follows. Definit le type NLMPCC : min f(x) s.t. l <= x <= u lcon(tb) <= cnl(x) <= ucon

with

cnl(x) := c(x),G(x),H(x),G(x).*H(x)

MPCC.MPCCAtX — Type

Type: MPCCAtX Methods: update!, reinit!

MPCCAtX contains the information concerning a nonlinear problem at the iteration x. min{x ∈ ℜⁿ} f(x) subject to lcon <= c(x) <= ucon, lvar <= x <= uvar, lG <= G(x) _| H(x) >= lH.

Basic information is:

x : the current candidate for solution to our original problem

fx : which is the funciton evaluation at x

gx : which is the gradient evaluation at x

Hx : which is the hessian representation at x

mu : Lagrange multiplier of the bounds constraints

cx : evaluation of the constraint function at x

Jx : jacobian matrix of the constraint function at x

lambda : Lagrange multiplier of the constraints

current_time : time

evals : number of evaluations of the function

Note: * by default, unknown entries are set to nothing (except evals). * All these information (except for x and lambda) are optionnal and need to be update when required. The update is done trhough the update! function. * x and lambda are mandatory entries. If no constraints lambda = []. * The constructor check the size of the entries.

MPCC.MPCCStopping — Type

Type: MPCCStopping (specialization of GenericStopping) Methods: start!, stop!, updateandstart!, updateandstop!, fill_in!, reinit!, status

Stopping structure for non-linear programming problems using NLPModels. Input : - pb : an AbstractMPCCModel - state : The information relative to the problem, see GenericState - (opt) meta : Metadata relative to stopping criterion, see StoppingMeta. - (opt) mainstp : Stopping of the main loop in case we consider a Stopping of a subproblem. If not a subproblem, then nothing. - (opt) listofstates : ListStates designed to store the history of States. - (opt) userspecific_struct : Contains any structure designed by the user.

MPCCStopping(:: AbstractMPCCModel, :: AbstractState; meta :: AbstractStoppingMeta = StoppingMeta(), main_stp :: Union{AbstractStopping, Nothing} = nothing, list :: Union{ListStates, Nothing} = nothing, user_specific_struct :: Any = nothing, kwargs...)

Note:

optimality_check : takes two inputs (AbstractMPCCModel, MPCCAtX)

and returns a Float64 to be compared at 0.

designed for MPCCAtX State. Constructor checks that the State has the

required entries.

Warning:

optimality_check does not necessarily fill in the State.

General Functions

MPCC.WStat — Function

W-stationary check: WStat: verifies the Weak-stationary conditions required: state.gx

if bounds: state.mu

if constraints: state.cx, state.Jx, state.lambda

if complementarity: state.cGx, state.JGx, state.lambdaG state.cHx, state.JHx, state.lambdaH

MPCC.CStat — Function

C-stationary check: Compute the score corresponding to: a) W-stationarity b) the correct sign of multipliers for indices I_00 by checking: lambdaG .* lambdaH => 0 (i.e. max.(- state.lambdaG .* state.lambdaH, 0.0))

required: state.gx

if bounds: state.mu

if constraints: state.cx, state.Jx, state.lambda

if complementarity: state.cGx, state.JGx, state.lambdaG state.cHx, state.JHx, state.lambdaH

MPCC.MStat — Function

M-stationary check: Compute the score corresponding to: a) W-stationarity b) the correct sign of multipliers for indices I_00 by checking: lambdaG .* lambdaH => 0 (i.e. max.(- state.lambdaG .* state.lambdaH, 0.0)) and max(lambdaG, lambdaH) => 0 (i.e. max.(- max.(state.lambdaG, state.lambdaH), 0.0))

required: state.gx

if bounds: state.mu

if constraints: state.cx, state.Jx, state.lambda

if complementarity: state.cGx, state.JGx, state.lambdaG state.cHx, state.JHx, state.lambdaH

MPCC.SStat — Function

S-stationary check: Compute the score corresponding to: a) W-stationarity b) the correct sign of multipliers for indices I00 = { i : 0.5(Gi(x^)+H_i(x^)) <= actif }

required: state.gx

if bounds: state.mu

if constraints: state.cx, state.Jx, state.lambda

if complementarity: state.cGx, state.JGx, state.lambdaG state.cHx, state.JHx, state.lambdaH