MPCC

Types

MPCC.AbstractMPCCModel — Type

Introduce the MPCC : min f(x) l <= x <= u lb <= c(x) <= ub 0 <= G(x) | H(x) >= 0

Missing docstring.

Missing docstring for MPCC.AbstractMPCCModelMeta. Check Documenter's build log for details.

MPCC.MPCCModelMeta — Type

A composite type that represents the main features of the optimization problem

optimize obj(x) subject to lvar ≤ x ≤ uvar lcon ≤ cons(x) ≤ ucon lcc ≤ G(x) | H(x) >= ucc

where x is an nvar-dimensional vector, obj is the real-valued objective function, cons is the vector-valued constraint function, optimize is either "minimize" or "maximize".

Here, lvar, uvar, lcon and ucon are vectors. Some of their components may be infinite to indicate that the corresponding bound or general constraint is not present.

Missing docstring.

Missing docstring for MPCC.MPCCCounters. Check Documenter's build log for details.

MPCC.ADMPCCModel — Type

ADMPCCModel is an AbstractMPCCModel using ForwardDiff to compute the derivatives. In this interface, the objective function $f$ and an initial estimate are required. If there are constraints, the function $c:ℝⁿ → ℝᵐ$ and the vectors $c_L$ and $c_U$ also need to be passed. Bounds on the variables and an inital estimate to the Lagrangian multipliers can also be provided.

ADMPCCModel(f, x0; lvar = [-∞,…,-∞], uvar = [∞,…,∞], y0 = zeros,
  c = NotImplemented, lcon = [-∞,…,-∞], ucon = [∞,…,∞], name = "Generic")

f - The objective function $f$. Should be callable;

x0 :: AbstractVector - The initial point of the problem;

lvar :: AbstractVector - $ℓ$, the lower bound of the variables;

uvar :: AbstractVector - $u$, the upper bound of the variables;

c - The constraints function $c$. Should be callable;

y0 :: AbstractVector - The initial value of the Lagrangian estimates;

lcon :: AbstractVector - $c_L$, the lower bounds of the constraints function;

ucon :: AbstractVector - $c_U$, the upper bounds of the constraints function;

name :: String - A name for the model.

The functions follow the same restrictions of ForwardDiff functions, summarised here:

The function can only be composed of generic Julia functions;

The function must accept only one argument;

The function's argument must accept a subtype of AbstractVector;

The function should be type-stable.

For contrained problems, the function $c$ is required, and it must return an array even when m = 1, and $c_L$ and $c_U$ should be passed, otherwise the problem is ill-formed. For equality constraints, the corresponding index of $c_L$ and $c_U$ should be the same.

MPCC.BPMPCCModel — Type

This type convert a bilevel problem as an MPCCModel.

Bilevel Program:

min_{x,y} f1(x,y)
s.t.  lcon    <= c(x)   <= ucon
      lvar    <= x      <= uvar
      y in argmin_{y} f2(x,y)
                 s.t. ll_lcon <= g(x,y) <= ll_ucon (l1,  l2)
                      ll_lvar <= y      <= ll_uvar (mu1, mu2)

MPCC reformulation of the Bilevel Program: min{x,y,l1,l2,mu1,mu2} f1(x,y) s.t. lvar <= x <= uvar lcon <= c(x) <= ucon 0 <= ∇y f2(x,y) - ∇y g(x,y)' * (l1 - l2) - (mu1 - mu2) <= 0 G(x):=( g(x,y) - lllcon) llucon - g(x,y) ) lluvar - y y - lllvar ) H(x):=( l1, l2, mu1, mu2 ) 0 <= G(x) _| H(x) >= 0 ```` Gradient, Jacobian and Hessian are computed using automatic differentiation

set nothirdderivative as true, to ignore the 3rd derivative of g.

MPCC.MPCCAmpl — Type

Definit le type MPCC : min f(x) l <= x <= u lb <= c(x) <= ub

with an array of integer "cvar" of the same size as c(x). cvar[i] is > 0 if ci(x) is involved in a complementarity constraint: (lbi = ci(x) or ci(x) = ub_i) OR x[cvar[i]] = 0.

Note: cons(nlp, x) call twice cons(nlp.mp, x) [TO FIX - idem jac, hess]

TODO: connect with https://github.com/tmigot/AmplNLReader.jl/blob/master/src/amplccmodel.jl

MPCC.MPCCNLPs — Type

Definit le type MPCC : min f(x) l <= x <= u lb <= c(x) <= ub lccG <= G(x) | H(x) >= lccH

MPCC.NLMPCC — Type

Convert an MPCCModel to an NLPModels as follows. Definit le type NLMPCC : min f(x) s.t. l <= x <= u lcon(tb) <= cnl(x) <= ucon

with

cnl(x) := c(x),G(x),H(x),G(x).*H(x)

MPCC.MPCCAtX — Type

Type: MPCCAtX Methods: update!, reinit!

MPCCAtX contains the information concerning a nonlinear problem at the iteration x. min{x ∈ ℜⁿ} f(x) subject to lcon <= c(x) <= ucon, lvar <= x <= uvar, lG <= G(x) _| H(x) >= lH.

Basic information is:

x : the current candidate for solution to our original problem

fx : which is the funciton evaluation at x

gx : which is the gradient evaluation at x

Hx : which is the hessian representation at x

mu : Lagrange multiplier of the bounds constraints

cx : evaluation of the constraint function at x

Jx : jacobian matrix of the constraint function at x

lambda : Lagrange multiplier of the constraints

current_time : time

evals : number of evaluations of the function

Note: * by default, unknown entries are set to nothing (except evals). * All these information (except for x and lambda) are optionnal and need to be update when required. The update is done trhough the update! function. * x and lambda are mandatory entries. If no constraints lambda = []. * The constructor check the size of the entries.

MPCC.MPCCStopping — Type

Type: MPCCStopping (specialization of GenericStopping) Methods: start!, stop!, updateandstart!, updateandstop!, fill_in!, reinit!, status

Stopping structure for non-linear programming problems using NLPModels. Input : - pb : an AbstractMPCCModel - optimalitycheck : a stopping criterion through an admissibility function - state : The information relative to the problem, see GenericState - maxcntrs : Dict contains the max number of evaluations - (opt) meta : Metadata relative to stopping criterion. - (opt) main_stp : Stopping of the main loop in case we consider a Stopping of a subproblem. If not a subproblem, then nothing.

Note:

optimality_check : takes two inputs (AbstractMPCCModel, MPCCAtX)

and returns a Float64 to be compared at 0.

designed for MPCCAtX State. Constructor checks that the State has the

required entries.

Warning:

optimality_check does not necessarily fill in the State.

General Functions

MPCC.WStat — Function

W-stationary check: WStat: verifies the Weak-stationary conditions required: state.gx

if bounds: state.mu

if constraints: state.cx, state.Jx, state.lambda

if complementarity: state.cGx, state.JGx, state.lambdaG state.cHx, state.JHx, state.lambdaH

MPCC.CStat — Function

C-stationary check: Compute the score corresponding to: a) W-stationarity b) the correct sign of multipliers for indices I_00 by checking: lambdaG .* lambdaH => 0 (i.e. max.(- state.lambdaG .* state.lambdaH, 0.0))

required: state.gx

if bounds: state.mu

if constraints: state.cx, state.Jx, state.lambda

if complementarity: state.cGx, state.JGx, state.lambdaG state.cHx, state.JHx, state.lambdaH

MPCC.MStat — Function

M-stationary check: Compute the score corresponding to: a) W-stationarity b) the correct sign of multipliers for indices I_00 by checking: lambdaG .* lambdaH => 0 (i.e. max.(- state.lambdaG .* state.lambdaH, 0.0)) and max(lambdaG, lambdaH) => 0 (i.e. max.(- max.(state.lambdaG, state.lambdaH), 0.0))

required: state.gx

if bounds: state.mu

if constraints: state.cx, state.Jx, state.lambda

if complementarity: state.cGx, state.JGx, state.lambdaG state.cHx, state.JHx, state.lambdaH

MPCC.SStat — Function

S-stationary check: Compute the score corresponding to: a) W-stationarity b) the correct sign of multipliers for indices I00 = { i : 0.5(Gi(x^)+H_i(x^)) <= actif }

required: state.gx

if bounds: state.mu

if constraints: state.cx, state.Jx, state.lambda

if complementarity: state.cGx, state.JGx, state.lambdaG state.cHx, state.JHx, state.lambdaH